+++ NEWS.md
@@ -1,7 +1,7 @@
 # consort 1.1.0
 
 -   Re-write most of the codes, there's some changes with the parameters.
--   Improved the alignment of the nodes.
+-   Improved the alignment of the nodes, no need to provide a coordinates.
 -   Now the plots will be drawn at the final stage.
 -   New function `build_grviz` and `build_grid`.
 -   Print the diagram with Shinny and HTML.